+++ packages/shared/src/index.ts
@@ -1,5 +1,3 @@
-
-
 export function copy(content: string, cb: Function) {
   const aux = document.createElement("input"); 
   aux.setAttribute("value", content);