+++ packages/vueui/scripts/build.ts
@@ -3,28 +3,24 @@ import path from "path";
 import { config } from "../vite.config.js";
 import { build, InlineConfig, defineConfig, UserConfig } from "vite";
 const buildAll = async () => {
-
-  // // 全量打包
-  // await build()
+  // 全量打包
+  await build();
 
   const baseOutDir = config.build.outDir;
   // 复制 Package.json 文件
   const packageJson = require("../package.json");
-  packageJson.main = "xiezijie-ui.umd.js";
   fs.outputFile(
     path.resolve(baseOutDir, `package.json`),
     JSON.stringify(packageJson, null, 2)
   );
 
-  const srcDir = path.resolve(__dirname, "../src/");
+  const srcDir = path.resolve(__dirname, "../src/components/");
   const componentsDir = fs.readdirSync(srcDir).filter((name) => {
     // 只要目录不要文件，且里面包含index.ts
     const componentDir = path.resolve(srcDir, name);
     const isDir = fs.lstatSync(componentDir).isDirectory();
     return isDir && fs.readdirSync(componentDir).includes("index.ts");
-  })
-
-  console.log(componentsDir)
+  });
 
   for (let name of componentsDir) {
     const outDir = path.resolve(baseOutDir, name);
@@ -44,15 +40,13 @@ const buildAll = async () => {
     fs.outputFile(
       path.resolve(outDir, `package.json`),
       `{
-        "name": "xiezijie-ui/${name}",
+        "name": "@xiezijie/vueui/${name}",
         "main": "index.umd.js",
         "module": "index.umd.js",
       }`,
       `utf-8`
     );
   }
-    
-  
 };
 
 buildAll();