+++ packages/vueui/src/index.ts
@@ -1,18 +1,12 @@
 import "tailwindcss/tailwind.css";
-// import { createApp } from 'vue'
-// // 从一个单文件组件中导入根组件
-// import App from './App.vue'
-
-// const app = createApp(App)
-// app.mount('#app')
-
-import XzjButton from "./Button/Button.vue";
-import XzjIcon from "./Icon/Icon.vue";
+import { App } from "vue";
+import { Button as XzjButton } from "./components/Button/index";
+import { Icon as XzjIcon } from "./components/Icon/index";
 
 export { XzjButton, XzjIcon };
 
 export default {
-  install(app) {
+  install(app: App) {
     app.component("xzj-button", XzjButton);
     app.component("xzj-icon", XzjIcon);
   },