+++ packages/vueui/CHANGELOG.md
@@ -1,7 +1,45 @@
-# xiezijie-ui
+# @xiezijie/vueui
 
-## 1.0.0
+## 0.0.8
 
-### Major Changes
+### Patch Changes
 
 -
+
+## 0.0.7
+
+### Patch Changes
+
+-
+
+## 0.0.6
+
+### Patch Changes
+
+-
+
+## 0.0.5
+
+### Patch Changes
+
+- ceshi
+
+## 0.0.4
+
+### Patch Changes
+
+-
+
+## 0.0.3
+
+### Patch Changes
+
+- 新增 App 支持
+
+## 0.0.2
+
+### Patch Changes
+
+- ceshi
+- Updated dependencies
+  - @xiezijie/shared@0.0.2